move logic process to service layer(modify post )
