+++ api/v1/post.py
@@ -1,6 +1,5 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-
 import service.post
 import repository.post
 from schema.database.postupdate import PostUpdate
@@ -27,20 +26,23 @@ def create_post(post: PostCreate, db: Session = Depends(get_db)):
 @router.delete("/{post_id}")
 def delete(post_id: int, db: Session = Depends(get_db)):
     post = service.post.get_post(db, post_id)
+    if post is None:
+        raise HTTPException(status_code=404, detail="Event not found")
     return service.post.delete(db=db, post=post)
 
 
 @router.patch("/{post_id}")
 def update_post(post_id: int, post_update: PostUpdate, db: Session = Depends(get_db)):
-    existing_post = service.post.get_post(db, post_id)
+    existing_post = service.post.get_post_by_id(db, post_id)
     if existing_post is None:
         raise HTTPException(status_code=404, detail="Post not found")
-    updated_post = service.post.patch_post(db, post_id, post_update)
+    updated_post = service.post.patch(db, post_id, post_update)
     return updated_post
+
 
 @router.get("/{post_id}")
 def get_post(post_id: int, db: Session = Depends(get_db)):
-    post = service.post.get_post(db, post_id)
+    post = service.post.get_post_by_id(db, post_id)
     if post is None:
         raise HTTPException(status_code=404, detail="Post not found")
     return post
