+++ service/post.py
@@ -1,25 +1,17 @@
-from database.event import Event
 from sqlalchemy.orm import Session
-
-from database.post import Post
 from schema.database.postupdate import PostUpdate
+from schema.database.post import Post
+from repository.post import patch_post, get_post
+from repository.post import delete as delete_post_repo
 
+def patch(db: Session, post_id: int, post_update: PostUpdate):
+    existing_post = get_post(db, post_id)
+    if existing_post is None:
+        return None
+    return patch_post(db, post_id, post_update)
 
-def get_post(db: Session, post_id: int):
-    return db.query(Post).filter(Post.id == post_id).first()
-
+def get_post_by_id(db: Session, post_id: int):
+    return get_post(db, post_id)
 
 def delete(db: Session, post: Post):
-    db.delete(post)
-    db.commit()
-
-
-def patch_post(db: Session, post_id: int, post_update: PostUpdate):
-    db_post = get_post(db, post_id)
-    if db_post:
-        for field, value in post_update.dict(exclude_unset=True).items():
-            setattr(db_post, field, value)
-        db.commit()
-        db.refresh(db_post)
-        return db_post
-    return None
+    delete_post_repo(db=db, post=post)
